+++ test/oap2/format/OAR042/ok-example.js
@@ -1,0 +1,9 @@
+module.exports = {
+  "swagger": "2.0",
+  "info": {
+    "version": "1.0.0",
+    "title": "Swagger Petstore"
+  },
+  "basePath": "/api-petstore/v3",
+  "paths": {}
+};